requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iso8601==1.1.0
passlib==1.7.4
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
